Move all authentication requests to the client
